requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.2
django-autoslug==1.9.8
django-heroku==0.3.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
Pillow==9.0.1
psycopg2==2.9.3
PyPDF2==1.26.0
python-bidi==0.4.2
razorpay==1.3.0
reportlab==3.6.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.5
